backend/services/gmail_service.py
# ╔══════════════════════════════════════════════════════════════════╗
# ║  VIKAA.AI — PUBLIC SHOWCASE                                      ║
# ║  This file requires environment credentials to run.              ║
# ║  Architecture and API signatures are shown for reference.        ║
# ║  See README.md and docs/ for setup guide.                        ║
# ╚══════════════════════════════════════════════════════════════════╝

"""
Gmail service — handles OAuth 2.0 and Gmail API reads.

Credential resolution order:
  Local dev  : credentials.json file (path from GMAIL_OAUTH_CREDENTIALS_PATH)
               tokens/<email>.json saved after each OAuth flow (multi-account)
               token.json legacy fallback (single-account)
  Render.com : Secret Files at /etc/secrets/token1.json, token2.json, ...
               GMAIL_ACCOUNTS env var = comma-separated email list (same order as token files)
               e.g. GMAIL_ACCOUNTS=debnath.kunal@gmail.com,story360degree@gmail.com
               Legacy: /etc/secrets/token.json (or token.json); GMAIL_DEFAULT_EMAIL optional —
               if unset, the service mailbox is resolved once via Gmail users.getProfile.

Until real credentials are provided every call returns DEMO_MODE = True
with mock data so the UI can be developed and tested independently.
"""

import os
import json
import logging
import base64
from datetime import datetime, timedelta, timezone
from typing import Optional

from google.auth.exceptions import RefreshError

logger = logging.getLogger(__name__)

TOKENS_DIR = "tokens"

# ── Credential helpers ─────────────────────────────────────────────────────────

SCOPES = [
    "https://www.googleapis.com/auth/gmail.readonly",
    "https://www.googleapis.com/auth/gmail.compose",
]


def _oauth_reauthorize_hint(account: Optional[str]) -> str:
    ...


def _render_token_path(index: int) -> str:
    """Return the Render Secret File path for account at position index (0-based)."""
    # index 0 → /etc/secrets/token1.json, index 1 → token2.json, etc.
    ...


def list_accounts() -> list[str]:
    """Return all Gmail accounts that have a saved token."""
    # ── Local dev: tokens/<email>.json files ──
    ...


def _credentials_available() -> bool:
    """Return True when OAuth app credentials AND at least one token exist."""
    ...


def _token_path_for(account: Optional[str]) -> str:
    """Resolve the token file path for a given account."""
    # ── Local dev: named token file ──
    ...


def _get_gmail_credentials(account: Optional[str] = None):
    """
    Build and return a valid google.oauth2.credentials.Credentials object.
    Raises RuntimeError when credentials are not yet configured.
    """
    ...


def _build_gmail(account: Optional[str] = None):
    """Return an authorised Gmail API service object."""
    ...


# Cached label for legacy single-token.json setups (avoids "default" mismatch vs JWT email in production).
_legacy_single_account_resolved: Optional[str] = None


def _gmail_me_email_safe(account: Optional[str] = None) -> Optional[str]:
    """Return mailbox address for this token via Gmail API; None on failure."""
    ...


def _resolved_legacy_single_account_label() -> str:
    """
    When only a single token file exists and GMAIL_DEFAULT_EMAIL is unset, list_accounts()
    used to return ['default'], so production users signing in with their real @gmail.com
    never matched. Resolve the actual mailbox from the token once and cache it.
    """
    ...


# ── Email fetching ─────────────────────────────────────────────────────────────

def _decode_body(payload: dict) -> str:
    """Extract plain-text body from a Gmail message payload."""
    ...


def _header(headers: list, name: str) -> str:
    ...


def fetch_emails(since_days: int = 1, max_results: int = 50, account: Optional[str] = None) -> list[dict]:
    """
    Return a list of simplified email dicts from the inbox.
    Falls back to DEMO data when credentials are not configured.
    """
    ...


# ── Demo / placeholder data ────────────────────────────────────────────────────

def _demo_emails(since_days: int) -> list[dict]:
    """Realistic-looking mock emails for UI development."""
    ...
