requirements.txt
# Core libraries
tavily-python
# Anthropic Claude SDK (MCP tool-use)
anthropic>=0.40.0
# True MCP server (gmail_mcp_server.py — stdio transport)
mcp>=1.0.0
# Gmail API + OAuth
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
fastapi==0.115.2
uvicorn==0.34.0
python-dotenv==1.0.1
dnspython==2.7.0 
email-validator==2.2.0
setuptools
setuptools-scm
supabase==2.15.2
# gotrue will be handled by supabase package

# Scheduler
APScheduler==3.10.4

# Document Processing
python-docx==1.1.2
python-pptx==1.0.2
PyPDF2==3.0.1
yt_dlp==2025.5.22

# Media Processing
moviepy==1.0.3
openpyxl==3.1.5
imageio[ffmpeg]==2.37.0
pillow==10.4.0
imageio_ffmpeg==0.6.0
ffmpeg-python==0.2.0
soundfile==0.13.1
openai-whisper

# Common dependencies
openai>=1.0.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic-core==2.33.1
anyio>=4.8.0,<5.0.0
tiktoken>=0.7,<1.0
tenacity==8.2.3
packaging==23.2
numpy==1.26.4
requests==2.32.3

# Other utilities
decorator==4.0.2
python-multipart==0.0.9
aiohttp==3.11.18
aiofiles==23.2.1
rich==13.7.0

# Testing and External APIs
pytest==8.3.2
pymongo>=4.6.1
tweepy==4.15.0
speedtest==0.0.1
GPUtil==1.4.0

# LangChain Ecosystem
langchain==0.3.25
langchain-core>=0.3.75
langchain-community==0.3.25
langchain-openai==0.2.8
langchain-mongodb==0.7.0
langchain-google-genai==2.1.8
langchain-experimental==0.3.4
google-genai==1.7.0

# Chat with Your Files / RAG Dependencies
sentence-transformers==2.2.2
pinecone-client==3.0.0
beautifulsoup4==4.13.4
pandas==2.3.1
duckduckgo-search>=6.3.0
mlflow>=2.14.0
ragas>=0.1.21
langsmith>=0.1.0
